+++ src/main/kotlin/com/github/warningimhack3r/npmupdatedependencies/backend/engine/NUDState.kt
@@ -2,12 +2,26 @@ package com.github.warningimhack3r.npmupdatedependencies.backend.engine
 
 import com.github.warningimhack3r.npmupdatedependencies.backend.data.Deprecation
 import com.github.warningimhack3r.npmupdatedependencies.backend.data.Versions
+import com.github.warningimhack3r.npmupdatedependencies.ui.statusbar.StatusBarHelper
 
-object NUDCache {
+object NUDState {
     val availableUpdates = mutableMapOf<String, Versions>()
     val deprecations = mutableMapOf<String, Deprecation>()
     val packageRegistries = mutableMapOf<String, String>()
 
     var isScanningForUpdates = false
+        set(value) {
+            field = value
+            StatusBarHelper.updateWidget()
+        }
     var isScanningForDeprecations = false
+        set(value) {
+            field = value
+            StatusBarHelper.updateWidget()
+        }
+    var isScanningForRegistries = false
+        set(value) {
+            field = value
+            StatusBarHelper.updateWidget()
+        }
 }